chore: `TUI_ORDERED_SHORT_WEEK_DAYS` add token factory
