+++ projects/addon-mobile/internal/primitive-calendar-mobile/primitive-calendar-mobile.component.ts
@@ -1,10 +1,6 @@
 import {ChangeDetectionStrategy, Component, HostBinding, Inject} from '@angular/core';
 import {TUI_IS_IOS} from '@taiga-ui/cdk';
-import {
-    TUI_ORDERED_SHORT_WEEK_DAYS,
-    TUI_PRIMITIVE_CALENDAR_PROVIDERS,
-    TuiPrimitiveCalendarComponent,
-} from '@taiga-ui/core';
+import {TUI_ORDERED_SHORT_WEEK_DAYS, TuiPrimitiveCalendarComponent} from '@taiga-ui/core';
 import {Observable} from 'rxjs';
 
 /**
@@ -15,7 +11,6 @@ import {Observable} from 'rxjs';
     selector: 'tui-primitive-calendar-mobile',
     templateUrl: './primitive-calendar-mobile.template.html',
     styleUrls: ['./primitive-calendar-mobile.style.less'],
-    providers: TUI_PRIMITIVE_CALENDAR_PROVIDERS,
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class TuiPrimitiveCalendarMobileComponent extends TuiPrimitiveCalendarComponent {
