+++ projects/core/components/primitive-calendar/index.ts
@@ -1,4 +1,4 @@
 export * from './primitive-calendar.component';
 export * from './primitive-calendar.module';
 export * from './pipes';
-export * from './primitive-calendar.providers';
+export * from './tokens';
