+++ projects/core/components/primitive-calendar/tokens/index.ts
@@ -1,0 +1,1 @@
+export * from './ordered-short-week-days';
